+++ packages/core/package.json
@@ -6,7 +6,6 @@
 	"types": "dist/index.d.ts",
 	"scripts": {
 		"test": "jest",
-		"dev": "nodemon -w src -i **.test.ts src/index.ts",
 		"clean": "rimraf ./dist",
 		"build": "pnpm clean && NODE_ENV=production tsc --project tsconfig.build.json",
 		"start": "NODE_ENV=production node ./dist/index.js"
