+++ packages/server/package.json
@@ -12,6 +12,7 @@
 	"dependencies": {
 		"@listening-party/app": "workspace:*",
 		"@listening-party/core": "workspace:*",
+		"body-parser": "^1.20.1",
 		"cors": "^2.8.5",
 		"cross-fetch": "^3.1.5",
 		"dotenv": "^16.0.3",
@@ -22,6 +23,7 @@
 		"ws": "^8.8.1"
 	},
 	"devDependencies": {
+		"@types/body-parser": "^1.19.2",
 		"@types/cors": "^2.8.12",
 		"@types/express": "^4.17.13",
 		"@types/jest": "^29.0.0",
