+++ packages/server/src/sockets/events/createRoom.ts
@@ -1,4 +1,5 @@
-import { rooms } from "../../createServer";
+import { rooms } from "../../models/room";
+
 import { sendToSocket } from "../sendToSocket";
 import { generateRandomString } from "../../utils/generateRandomString";
 
@@ -6,20 +7,20 @@ import type { WebSocket } from "ws";
 import type { SocketEventHandler } from "../../models/socket";
 
 const createRoom: SocketEventHandler<"CreateRoom"> = (userId, socket, data) => {
-	let roomId = "";
-	while (!roomId || rooms.has(roomId)) {
-		roomId = generateRandomString(6, "0123456789abcdefghijklmnopqrstuvwxyz");
-	}
+  let roomId = "";
+  while (!roomId || rooms.has(roomId)) {
+    roomId = generateRandomString(6, "0123456789abcdefghijklmnopqrstuvwxyz");
+  }
 
-	rooms.set(roomId, {
-		name: data.roomName,
-		hostId: userId,
-		connections: new Map<string, WebSocket>(),
-	});
+  rooms.set(roomId, {
+    name: data.roomName,
+    hostId: userId,
+    connections: new Map<string, WebSocket>(),
+  });
 
-	sendToSocket("CreateSuccessful", socket, {
-		roomId,
-	});
+  sendToSocket("CreateSuccessful", socket, {
+    roomId,
+  });
 };
 
 export { createRoom };
